backend/requirements.txt
fastapi
uvicorn
sqlalchemy
pymysql
langgraph
langchain-groq
pydantic
python-dotenv
